fastapi_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
requests==2.31.0
passlib==1.7.4
python-jose==3.3.0
cryptography==41.0.7
PyJWT
passlib
passlib[bcrypt]
psycopg2
psycopg2-binary
bcrypt
argon2-cffi
sendgrid